final app metrics implementation and looped axios plugin fix
