+++ nuxt.config.ts
@@ -32,6 +32,7 @@ export default defineNuxtConfig({
       ALTERNATIVE_NETWORK_URL: process.env.ALTERNATIVE_NETWORK_URL,
       ALTERNATIVE_NETWORK_NAME: process.env.ALTERNATIVE_NETWORK_NAME,
       AE_TOKEN_ID: process.env.AE_TOKEN_ID,
+      DEBUG_MODE: process.env.DEBUG_MODE,
     },
   },
   postcss: {
